Add-ons
I have created 2 empty game objects for the music being used for the Farm and Town scene. Also, I created and prefab Audio Manager object. Also, prefab the songs that were used in each of the scenes. I created another folder in Prefab. One is labeled Audio for the Audio Manager object and the songs are in the Music Prefab folder.